requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.1.8
dnspython==2.7.0
eventlet==0.36.1
exceptiongroup==1.3.0
fastapi==0.115.0
Flask==3.0.0
flask-sock==0.7.0
frozenlist==1.8.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
numpy==1.26.4
packaging==25.0
propcache==0.4.0
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.38.6
twilio==9.2.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.3
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.22.0
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
